Move command.go to common
